% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{condition_table}
\alias{condition_table}
\title{Condition Data}
\format{A data frame consisting of 340351 value and 3 variables:
 \code{CONDITION_ID}, \code{NCT_ID}, \code{CONDITION}}
\usage{
condition_table
}
\description{
Contains data about trial primary condition / disease (per NCT_ID)
}
\keyword{datasets}

