% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{facilities_table}
\alias{facilities_table}
\title{Location Data}
\format{A data frame consisting of 1633395 value and 8 variables:
 \code{FACILITY_ID}, \code{NCT_ID}, \code{STATUS}, \code{FACILITY_NAME},
 \code{CITY}, \code{STATE}, \code{ZIP}, \code{COUNTRY}}
\usage{
facilities_table
}
\description{
Contains data about trial location (per NCT_ID)
}
\keyword{datasets}

