% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clinical_trial_data.R
\name{get_clinical_trial_data}
\alias{get_clinical_trial_data}
\title{Get furnished clinical trial info}
\format{A data frame with 12 variables:
 \describe {
   \item{\code{NCT_ID}}{Unique id for every trial}
   \item{\code{BRIEF_TITLE}}{Brief trial title}
   \item{\code{OFFICIAL_TITLE}}{Actual trial title}
   \item{\code{SOURCE}}{Similar to Lead sponsor, but may indicate who the data
                        was submitted on behalf of, or a subsidiary of the sponsor.
                        }
  \item{\code{OVERALL_STATUS}}{Active, not recruiting,
                             Approved for marketing,
                             Available,
                             Completed,
                             Enrolling by invitation,
                             No longer available,
                             Not yet recruiting,
                             Recruiting,
                             Suspended,
                             Temporarily not available,
                             Terminated,
                             Withdrawn,
                             Withheld
                             }
   \item{\code{START_DATE}}{Date in MM YYYY format eg. Jan 2012}
   \item{\code{COMPLETION_DAYE}}{Date in MM YYYY format eg. Jan 2012}
   \item{\code{PHASE}}{Phase 0,
                    Phase 1,
                    Phase 2,
                    Phase 1/Phase 2,
                    Phase 3,
                    Phase 2/Phase 3,
                    Phase 4
                    N/A
                    }
   \item{\code{ENROLLMENT}}{number of participants in the study in integer format}
   \item{\code{WHY_STOPPED}}{if trial suspended then the reason for suspension}
   \item{\code{AGENCY}}{Lead sponsor}
   \item{\code{CONDITION}}{Primary disease or condition being studied or focus of the study}
 }}
\usage{
get_clinical_trial_data()
}
\value{
A list consisting of 211437 values and  12 variables:
}
\description{
This function uses the raw clinical trial data and processes to only get relevant
information
}
\keyword{Clinical}
\keyword{data}
\keyword{trials}

