% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_k_values.R
\name{get_k_value}
\alias{get_k_value}
\title{Get K piecewise function based on kterm and location}
\usage{
get_k_value(location, kterm = "", start_date, end_date, phase = "")
}
\arguments{
\item{location}{Facility city}

\item{kterm}{Disease or primary condition under observation. Default is blank}

\item{start_date}{e.g. Jan 2012}

\item{end_date}{e.g. March 2013}
}
\value{
A list containing count of trials in operation per month between start and end date
}
\description{
Get K piecewise function based on kterm and location
}

