% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_matching_kterm_nct_id.R
\name{get_matching_kterm_nct_id}
\alias{get_matching_kterm_nct_id}
\title{Get matching nct_id for a given disease or primary condition}
\usage{
get_matching_kterm_nct_id(kterm)
}
\arguments{
\item{kterm}{A string representing a disease name or primary condition}
}
\value{
A list of matching NCT_ID
}
\description{
Get matching nct_id for a given disease or primary condition
}

