% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_k_values.R
\name{plot_k_values}
\alias{plot_k_values}
\title{Plot k values as a step function w.r.t time}
\usage{
plot_k_values(get_k_value_op, start_date, end_date, kterm = "",
  location = "")
}
\arguments{
\item{get_k_value_op}{Output from get_k_value()}

\item{start_date}{e.g. Jan 2012}

\item{end_date}{e.g. March 2013}

\item{kterm}{Disease / Primary condition (Optional)}

\item{location}{Optional}
}
\value{
Plot a step function
}
\description{
Plot k values as a step function w.r.t time
}

