% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_functions.R
\name{read_condition_data}
\alias{read_condition_data}
\title{Read Clinical Trial Condition Info}
\usage{
read_condition_data()
}
\arguments{
\item{None}{}
}
\description{
Returns a data frame consisting of 340351 value and 3 variables.
}

