% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_functions.R
\name{read_location_data}
\alias{read_location_data}
\title{Read Clinical Trial Location Info}
\usage{
read_location_data()
}
\arguments{
\item{None}{}
}
\description{
Return a data frame consisting of 1633395 value and 8 variables:
}

