% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_functions.R
\name{read_raw_clinical_data}
\alias{read_raw_clinical_data}
\title{Read Raw Clinical Trial Info}
\usage{
read_raw_clinical_data()
}
\arguments{
\item{None}{}
}
\description{
Returns a data frame consisting of 211437 values and  42 variables:
}

