% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_functions.R
\name{read_sponsor_data}
\alias{read_sponsor_data}
\title{Read Clinical Trial Sponsor Info}
\usage{
read_sponsor_data()
}
\arguments{
\item{None}{}
}
\description{
Returns a data frame consisting of 328709 value and 5 variables:.
}

