% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_clinical_trial.R
\name{search_clinical_trial}
\alias{search_clinical_trial}
\title{Get clinical trial data by matching specific NCT_ID for the specific kterm}
\usage{
search_clinical_trial(kterm)
}
\arguments{
\item{kterm}{A string representing a disease name or primary condition}
}
\description{
Get clinical trial data by matching specific NCT_ID for the specific kterm
}

