% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_locations.R
\name{search_locations}
\alias{search_locations}
\title{Get clinical trial location data by matching specific NCT_ID for the specific kterm}
\usage{
search_locations(kterm)
}
\arguments{
\item{kterm}{A string representing a disease name or primary condition}
}
\value{
A dataframe containig location of facilities where trials are conducted
}
\description{
Get clinical trial location data by matching specific NCT_ID for the specific kterm
}

