% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sponsor_table}
\alias{sponsor_table}
\title{Sponsor Data}
\format{A data frame consisting of 328709 value and 5 variables:
 \code{SPONSOR_ID}, \code{NCT_ID}, \code{SPONSOR_TYPE}, \code{AGENCY},
 \code{AGENCY_CLASS}}
\usage{
sponsor_table
}
\description{
Contains data about sponsors (per NCT_ID)
}
\keyword{datasets}

